Layout enhancement on Home * Projects section
The language text is now color coded and bordered, giving it an icon type of feel. I used Pixlr to add cell phones to the desktops in the projects section to display mobile versions of each project. I gave the home page an image with an Ipad and two cell-phones to display the mobile responsiveness of each through a picture.